Refactor JiraIssue to GraphQL SDL first
