+++ packages/server/graphql/public/typeDefs/JiraIssue.graphql
@@ -1,0 +1,107 @@
+"""
+The Jira Issue that comes direct from Jira
+"""
+type JiraIssue implements TaskIntegration {
+  """
+  GUID cloudId:issueKey
+  """
+  id: ID!
+
+  """
+  The parabol teamId this issue was fetched for
+  """
+  teamId: ID!
+
+  """
+  The parabol userId this issue was fetched for
+  """
+  userId: ID!
+
+  """
+  The ID of the jira cloud where the issue lives
+  """
+  cloudId: ID!
+
+  """
+  The name of the jira cloud where the issue lives
+  """
+  cloudName: ID!
+
+  """
+  The url to access the issue
+  """
+  url: URL!
+
+  """
+  The key of the issue as found in Jira
+  """
+  issueKey: ID!
+
+  """
+  The type of the issue
+  """
+  issueType: ID!
+
+  """
+  The key of the project, which is the prefix to the issueKey
+  """
+  projectKey: ID!
+
+  """
+  The project fetched from jira
+  """
+  project: JiraRemoteProject
+
+  """
+  The plaintext summary of the jira issue
+  """
+  summary: String!
+
+  """
+  Field names that exists on the issue and can be used as estimation fields
+  """
+  possibleEstimationFieldNames: [String!]!
+
+  """
+  The stringified ADF of the jira issue description
+  """
+  description: String!
+
+  """
+  The description converted into raw HTML
+  """
+  descriptionHTML: String!
+}
+
+"""
+A connection to a list of items.
+"""
+type JiraIssueConnection {
+  """
+  Page info with cursors coerced to ISO8601 dates
+  """
+  pageInfo: PageInfoDateCursor
+
+  """
+  A list of edges.
+  """
+  edges: [JiraIssueEdge!]!
+
+  """
+  An error with the connection, if any
+  """
+  error: StandardMutationError
+}
+
+"""
+An edge in a connection.
+"""
+type JiraIssueEdge {
+  """
+  The item at the end of the edge
+  """
+  node: JiraIssue!
+  cursor: DateTime
+}
+
+extend union NotificationSubscriptionPayload = JiraIssue
