+++ packages/server/graphql/public/typeDefs/_legacy.graphql
@@ -1491,29 +1491,6 @@ type AtlassianIntegration {
   userId: ID!
 
   """
-  A list of issues coming straight from the jira integration for a specific team member
-  """
-  issues(
-    first: Int = 100
-
-    """
-    the datetime cursor
-    """
-    after: DateTime
-
-    """
-    A string of text to search for, or JQL if isJQL is true
-    """
-    queryString: String
-
-    """
-    true if the queryString is JQL, else false
-    """
-    isJQL: Boolean!
-    projectKeyFilters: [ID!]
-  ): JiraIssueConnection!
-
-  """
   A list of projects accessible by this team member. empty if viewer is not the user
   """
   projects: [JiraRemoteProject!]!
@@ -1532,107 +1509,6 @@ type AtlassianIntegration {
   the list of suggested search queries, sorted by most recent. Guaranteed to be < 60 days old
   """
   jiraSearchQueries: [JiraSearchQuery!]!
-}
-
-"""
-A connection to a list of items.
-"""
-type JiraIssueConnection {
-  """
-  Page info with cursors coerced to ISO8601 dates
-  """
-  pageInfo: PageInfoDateCursor
-
-  """
-  A list of edges.
-  """
-  edges: [JiraIssueEdge!]!
-
-  """
-  An error with the connection, if any
-  """
-  error: StandardMutationError
-}
-
-"""
-An edge in a connection.
-"""
-type JiraIssueEdge {
-  """
-  The item at the end of the edge
-  """
-  node: JiraIssue!
-  cursor: DateTime
-}
-
-"""
-The Jira Issue that comes direct from Jira
-"""
-type JiraIssue implements TaskIntegration {
-  """
-  GUID cloudId:issueKey
-  """
-  id: ID!
-
-  """
-  The parabol teamId this issue was fetched for
-  """
-  teamId: ID!
-
-  """
-  The parabol userId this issue was fetched for
-  """
-  userId: ID!
-
-  """
-  The ID of the jira cloud where the issue lives
-  """
-  cloudId: ID!
-
-  """
-  The name of the jira cloud where the issue lives
-  """
-  cloudName: ID!
-
-  """
-  The url to access the issue
-  """
-  url: URL!
-
-  """
-  The key of the issue as found in Jira
-  """
-  issueKey: ID!
-
-  """
-  The key of the project, which is the prefix to the issueKey
-  """
-  projectKey: ID!
-
-  """
-  The project fetched from jira
-  """
-  project: JiraRemoteProject
-
-  """
-  The plaintext summary of the jira issue
-  """
-  summary: String!
-
-  """
-  Field names that exists on the issue and can be used as estimation fields
-  """
-  possibleEstimationFieldNames: [String!]!
-
-  """
-  The stringified ADF of the jira issue description
-  """
-  description: String!
-
-  """
-  The description converted into raw HTML
-  """
-  descriptionHTML: String!
 }
 
 interface TaskIntegration {
@@ -10846,7 +10722,6 @@ union NotificationSubscriptionPayload =
   | User
   | AuthTokenPayload
   | PersistGitHubSearchQuerySuccess
-  | JiraIssue
 
 type AddNewFeaturePayload {
   """
